+++ lib/src/alert_dialog/alert_dialog.dart
@@ -1,5 +1,6 @@
 export 'alert_dialog_action.dart';
 export 'show_alert_dialog.dart';
+export 'show_confirmatoin_dialog.dart';
 export 'show_dialog.dart';
 export 'show_ok_alert_dialog.dart';
 export 'show_ok_cancel_alert_dialog.dart';