+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.6.0
+
+- Add `showConfirmationDialog`
+  - Show [confirmation dialog](https://material.io/components/dialogs#confirmation-dialog), whose appearance is adaptive according to platform
+  - For cupertino, fallback to ActionSheet
+
 ## 0.5.0
 
 - Add `materialConfiguration` to `showModalBottomSheet`